refactor: new Recovery facet
