+++ packages/ats/contracts/contracts/facets/layer_1/ERC3643/ERC3643Management.sol
@@ -1,38 +1,13 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity >=0.8.0 <0.9.0;
 
-// solhint-disable max-line-length
-
-import { AGENT_ROLE } from "../../../constants/roles.sol";
 import { IERC3643Management } from "./IERC3643Management.sol";
 import { Modifiers } from "../../../services/Modifiers.sol";
 import { ERC3643StorageWrapper } from "../../../domain/core/ERC3643StorageWrapper.sol";
-import { TimeTravelStorageWrapper } from "../../../test/testTimeTravel/timeTravel/TimeTravelStorageWrapper.sol";
 
 abstract contract ERC3643Management is IERC3643Management, Modifiers {
     // solhint-disable-next-line func-name-mixedcase
     function initialize_ERC3643(address _compliance, address _identityRegistry) external onlyNotERC3643Initialized {
         ERC3643StorageWrapper.initialize_ERC3643(_compliance, _identityRegistry);
-    }
-
-    function recoveryAddress(
-        address _lostWallet,
-        address _newWallet,
-        address _investorOnchainID
-    )
-        external
-        override
-        onlyRole(AGENT_ROLE)
-        onlyUnrecoveredAddress(_lostWallet)
-        onlyEmptyWallet(_lostWallet)
-        onlyWithoutMultiPartition
-        returns (bool success_)
-    {
-        success_ = ERC3643StorageWrapper.recoveryAddress(
-            _lostWallet,
-            _newWallet,
-            _investorOnchainID,
-            TimeTravelStorageWrapper.getBlockTimestamp()
-        );
     }
 }
