+++ packages/ats/contracts/contracts/facets/layer_1/ERC3643/ERC3643ManagementFacet.sol
@@ -12,10 +12,9 @@ contract ERC3643ManagementFacet is ERC3643Management, IStaticFunctionSelectors {
     }
 
     function getStaticFunctionSelectors() external pure override returns (bytes4[] memory staticFunctionSelectors_) {
-        uint256 selectorIndex = 2;
+        uint256 selectorIndex = 1;
         staticFunctionSelectors_ = new bytes4[](selectorIndex);
         unchecked {
-            staticFunctionSelectors_[--selectorIndex] = this.recoveryAddress.selector;
             staticFunctionSelectors_[--selectorIndex] = this.initialize_ERC3643.selector;
         }
     }
