+++ packages/ats/contracts/contracts/facets/layer_1/ERC3643/IERC3643.sol
@@ -3,7 +3,6 @@ pragma solidity >=0.8.0 <0.9.0;
 
 import { IERC3643Types } from "./IERC3643Types.sol";
 import { IERC3643Management } from "./IERC3643Management.sol";
-import { IERC3643Read } from "./IERC3643Read.sol";
 
 /**
  * @title IERC3643
@@ -13,4 +12,4 @@ import { IERC3643Read } from "./IERC3643Read.sol";
  *      a concrete facet forces that facet to implement functions it does not own.
  */
 // solhint-disable no-empty-blocks
-interface IERC3643 is IERC3643Types, IERC3643Management, IERC3643Read {}
+interface IERC3643 is IERC3643Types, IERC3643Management {}
