--- packages/ats/contracts/contracts/test/testTimeTravel/facetsTimeTravel/eRC3643Read/ERC3643ReadFacetTimeTravel.sol
@@ -1,11 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Contract copy-pasted form OZ and extended
-
-pragma solidity >=0.8.0 <0.9.0;
-
-import { ERC3643ReadFacet } from "../../../../facets/layer_1/ERC3643/ERC3643ReadFacet.sol";
-import { TimeTravelProvider } from "../../timeTravel/TimeTravelProvider.sol";
-
-contract ERC3643ReadFacetTimeTravel is ERC3643ReadFacet, TimeTravelProvider {
-    // solhint-disable-next-line no-empty-blocks
-}
