+++ packages/ats/contracts/scripts/domain/bond/createConfiguration.ts
@@ -96,7 +96,7 @@ const BOND_FACETS = [
   "BatchMintFacet",
   "BatchTransferFacet",
   "ERC3643ManagementFacet",
-  "ERC3643ReadFacet",
+  "RecoveryFacet",
   "IdentityFacet",
   "ComplianceFacet",
   "ComplianceByPartitionFacet",
