+++ packages/ats/contracts/scripts/domain/bondFixedRate/createConfiguration.ts
@@ -91,7 +91,7 @@ const BOND_FIXED_RATE_FACETS = [
   "BatchMintFacet",
   "BatchTransferFacet",
   "ERC3643ManagementFacet",
-  "ERC3643ReadFacet",
+  "RecoveryFacet",
   "IdentityFacet",
   "ComplianceFacet",
   "ComplianceByPartitionFacet",
