+++ packages/ats/contracts/scripts/domain/equity/createConfiguration.ts
@@ -87,7 +87,7 @@ const EQUITY_FACETS = [
   "BatchMintFacet",
   "BatchTransferFacet",
   "ERC3643ManagementFacet",
-  "ERC3643ReadFacet",
+  "RecoveryFacet",
   "IdentityFacet",
   "ComplianceFacet",
   "ComplianceByPartitionFacet",
