+++ packages/ats/contracts/scripts/domain/loanPortfolio/createConfiguration.ts
@@ -94,7 +94,7 @@ const LOANS_PORTFOLIO_FACETS = [
   "BatchMintFacet",
   "BatchTransferFacet",
   "ERC3643ManagementFacet",
-  "ERC3643ReadFacet",
+  "RecoveryFacet",
   "IdentityFacet",
   "ComplianceFacet",
   "ComplianceByPartitionFacet",
